+++ CHANGELOG.md
@@ -7,7 +7,11 @@
 * None.
 
 ### Fixed
+* User exceptions now propagate correctly out from `RealmMigration` and `CompactOnLaunchCallback` instead of just resulting in a generic *User-provided callback failed* `RuntimeException`. (Issue [#1228](https://github.com/realm/realm-kotlin/issues/1228))
 * The default compact-on-launch callback trigger 50% or more of the space could be reclaimed was reversed. (Issue [#1380](https://github.com/realm/realm-kotlin/issues/1380))
+
+### Fixed
+* None.
 
 ### Compatibility
 * File format: Generates Realms with file format v23.